refactor: squareImage함수 extension으로 분리
