+++ Expo1900/Expo1900/Controllers/KoreaItemsViewController.swift
@@ -50,15 +50,14 @@ extension KoreaItemsViewController: UITableViewDataSource {
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: self.cellIdentifier) else {
+        let item = expositionItems[indexPath.row]
+        guard let cell = tableView.dequeueReusableCell(withIdentifier: self.cellIdentifier),
+                let image = UIImage(named: item.imageName)  else {
             return UITableViewCell()
         }
         
-        let item = expositionItems[indexPath.row]
         var content = cell.defaultContentConfiguration()
-        let image = UIImage(named: item.imageName)
-        
-        content.image = squreImage(at: image)
+        content.image = image.squareImage()
         content.imageToTextPadding = CGFloat(5)
         content.text = item.name
         content.secondaryText = item.shortDescription
@@ -68,42 +67,9 @@ extension KoreaItemsViewController: UITableViewDataSource {
         
         return cell
     }
-    
-    private func squreImage(at image: UIImage?, length: CGFloat = 60) -> UIImage? {
-        guard let image = image else { return image }
-        
-        let originWidth: CGFloat = image.size.width
-        let originHeigth: CGFloat = image.size.height
-        var resizedWidth: CGFloat = length
-        var resizedHeight: CGFloat = length
-        
-        UIGraphicsBeginImageContext(CGSize(width: length, height: length))
-        UIColor.clear.set()
-        UIRectFill(CGRect(x: 0.0, y: 0.0, width: length, height: length))
-        
-        let sizeRatio = length / max(originWidth, originHeigth)
-        if originWidth > originHeigth {
-            resizedWidth = length
-            resizedHeight = originHeigth * sizeRatio
-        } else {
-            resizedWidth = originWidth * sizeRatio
-            resizedHeight = length
-        }
-        
-        image.draw(in: CGRect(x: length / 2 - resizedWidth / 2,
-                              y: length / 2 - resizedHeight / 2,
-                              width: resizedWidth,
-                              height: resizedHeight))
-        
-        let resizedImage = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-        
-        return resizedImage
-    }
 }
 
 extension KoreaItemsViewController: UITableViewDelegate {
-    
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         tableView.deselectRow(at: indexPath, animated: true)
     }
